# api/api.swagger.yaml
openapi: 3.0.1
info:
  title: Pipeliner API
  contact: { }
  version: "0.1"
servers:
  - url: //localhost:8181/api/pipeliner/v1
paths:
...
  /application/rate/{workNumber}:
    post:
      summary: rate application
      parameters:
        - name: workNumber
          in: path
          description: work number
          required: true
          schema:
            type: string
      operationId: rateApplication
      requestBody:
        description: rate application
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RateApplicationRequest'
        required: true
      responses:
        200:
          description: OK
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpError'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpError'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpError'
...
components:
  schemas:
...
    RateApplicationRequest:
      type: object
      properties:
        comment:
          type: string
        rate:
          type: integer
...
    httpError:
      type: object
      properties:
        description:
          type: string
        error:
          type: string
        status_code:
          type: integer
...